3rd phase done 14-jan-2026 #3rd-Phase

+++ lib/presentation/invoice_pdf_generation_screen/model/invoice_pdf_preview_model.dart
@@ -135,7 +135,7 @@ class ReceiptLines {
   String? receiptLineName;
   int? receiptLineQuantity;
   double? receiptLineTotal;
-  int? taxPercent;
+  double? taxPercent;
   String? taxID;
 
   ReceiptLines(
